pyproject.toml
[project]
name = "odoo-like-crud"
version = "0.2.0"
description = "CRUD Odoo-like con Clean Architecture, FastAPI y MySQL"
requires-python = ">=3.11"
dependencies = [
  "PyMySQL>=1.1.0",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["servidor/tests"]
pythonpath = ["servidor", "cliente", "."]
markers = [
  "integration: tests de integraciÃ³n (requiere DB configurada)"
]

